finished-up shrinkwrap.py
